pyproject.toml
[tool.black]
exclude = "/(.venv)"
line-length = 79
target-version = ["py39"]

[tool.isort]
# https://pycqa.github.io/isort/docs/configuration/config_files.html
combine_as_imports = true
force_grid_wrap = 2
include_trailing_comma = true
lines_after_imports = -1
lines_between_types = 1
profile = "black"
skip_gitignore = true
skip_glob = [".venv/*", "maps/*"]
use_parentheses = true

[tool.autoflake]
exclude = [".venv", "maps/*"]
in-place = true
quiet = true
recursive = true
remove-all-unused-imports = true
remove-unused-variables = true


[project] 
name = "checklists"
dependencies = [
    "blinker==1.9.0",
    "click==8.1.8",
    "Flask==3.1.0",
    "itsdangerous==2.2.0",
    "Jinja2==3.1.6",
    "MarkupSafe==3.0.2",
    "Werkzeug==3.1.3",
    "recursive-diff",
    "PyYaml", 
    "peewee",
    "inject",
    "pydantic",
    "psycopg2"

]


version = "0.0.1"